requirements.txt
tensorflow-gpu==1.8.0
keras==2.2.0
h5py==2.8.0
tqdm==4.23.4
numpy==1.15.4